Restructure contour_statistics.
